Rename filtrer-tree to filter-tree
Fix french typo.
