+++ packages/react-components/src/components/xy-output-component.tsx
@@ -9,10 +9,10 @@ import { Entry } from 'tsp-typescript-client/lib/models/entry';
 import { ResponseStatus } from 'tsp-typescript-client/lib/models/response/responses';
 import { XYSeries } from 'tsp-typescript-client/lib/models/xy';
 import Chart = require('chart.js');
-import { EntryTree } from './utils/filtrer-tree/entry-tree';
-import { getAllExpandedNodeIds } from './utils/filtrer-tree/utils';
-import { TreeNode } from './utils/filtrer-tree/tree-node';
-import ColumnHeader from './utils/filtrer-tree/column-header';
+import { EntryTree } from './utils/filter-tree/entry-tree';
+import { getAllExpandedNodeIds } from './utils/filter-tree/utils';
+import { TreeNode } from './utils/filter-tree/tree-node';
+import ColumnHeader from './utils/filter-tree/column-header';
 import { signalManager, Signals } from 'traceviewer-base/lib/signals/signal-manager';
 import { BIMath } from 'timeline-chart/lib/bigint-utils';
 import { scaleLinear } from 'd3-scale';
